Move IndexedStack to indexed_stack.dart (previously visibility.dart)
