+++ packages/flutter/lib/src/widgets/indexed_stack.dart
@@ -9,14 +9,199 @@
 /// @docImport 'transitions.dart';
 library;
 
-import 'package:flutter/foundation.dart' show internal;
 import 'package:flutter/rendering.dart';
 
 import 'basic.dart';
+import 'debug.dart';
 import 'focus_scope.dart';
 import 'framework.dart';
 import 'sliver.dart';
 import 'ticker_provider.dart';
+
+/// A [Stack] that shows a single child from a list of children.
+///
+/// The displayed child is the one with the given [index]. The stack is
+/// always as big as the largest child.
+///
+/// If value is null, then nothing is displayed.
+///
+/// {@youtube 560 315 https://www.youtube.com/watch?v=_O0PPD1Xfbk}
+///
+/// {@tool dartpad}
+/// This example shows a [IndexedStack] widget being used to lay out one card
+/// at a time from a series of cards, each keeping their respective states.
+///
+/// ** See code in examples/api/lib/widgets/basic/indexed_stack.0.dart **
+/// {@end-tool}
+///
+/// See also:
+///
+///  * [Stack], for more details about stacks.
+///  * The [catalog of layout widgets](https://flutter.dev/widgets/layout/).
+class IndexedStack extends StatelessWidget {
+  /// Creates a [Stack] widget that paints a single child.
+  const IndexedStack({
+    super.key,
+    this.alignment = AlignmentDirectional.topStart,
+    this.textDirection,
+    this.clipBehavior = Clip.hardEdge,
+    this.sizing = StackFit.loose,
+    this.index = 0,
+    this.children = const <Widget>[],
+  });
+
+  /// How to align the non-positioned and partially-positioned children in the
+  /// stack.
+  ///
+  /// Defaults to [AlignmentDirectional.topStart].
+  ///
+  /// See [Stack.alignment] for more information.
+  final AlignmentGeometry alignment;
+
+  /// The text direction with which to resolve [alignment].
+  ///
+  /// Defaults to the ambient [Directionality].
+  final TextDirection? textDirection;
+
+  /// {@macro flutter.material.Material.clipBehavior}
+  ///
+  /// Defaults to [Clip.hardEdge].
+  final Clip clipBehavior;
+
+  /// How to size the non-positioned children in the stack.
+  ///
+  /// Defaults to [StackFit.loose].
+  ///
+  /// See [Stack.fit] for more information.
+  final StackFit sizing;
+
+  /// The index of the child to show.
+  ///
+  /// If this is null, none of the children will be shown.
+  final int? index;
+
+  /// The child widgets of the stack.
+  ///
+  /// Only the child at index [index] will be shown.
+  ///
+  /// See [Stack.children] for more information.
+  final List<Widget> children;
+
+  @override
+  Widget build(BuildContext context) {
+    // Each child is wrapped with _VisibilityScope (so Visibility.of reports
+    // the child as hidden when it is not the selected index) and with
+    // ExcludeFocus (so non-selected children cannot receive focus). Both of
+    // these do not introduce any RenderObjects between the child and the
+    // enclosing RenderIndexedStack.
+    //
+    // This allows ParentDataWidgets such as Positioned to correctly apply
+    // their StackParentData to the IndexedStack's render object.
+    //
+    // Painting, hit-testing, and semantics for non-selected children are
+    // already handled by RenderIndexedStack, so no additional render-object
+    // wrappers are needed.
+    final wrappedChildren = List<Widget>.generate(children.length, (int i) {
+      final isSelected = i == index;
+      return _VisibilityScope(
+        isVisible: isSelected,
+        child: ExcludeFocus(excluding: !isSelected, child: children[i]),
+      );
+    });
+    return _RawIndexedStack(
+      alignment: alignment,
+      textDirection: textDirection,
+      clipBehavior: clipBehavior,
+      sizing: sizing,
+      index: index,
+      children: wrappedChildren,
+    );
+  }
+}
+
+/// The render object widget that backs [IndexedStack].
+class _RawIndexedStack extends Stack {
+  /// Creates a [Stack] widget that paints a single child.
+  const _RawIndexedStack({
+    super.alignment,
+    super.textDirection,
+    super.clipBehavior,
+    StackFit sizing = StackFit.loose,
+    this.index = 0,
+    super.children,
+  }) : assert(
+         index == null ||
+             (index == 0 && children.length == 0) ||
+             (index >= 0 && index < children.length),
+         'The index must be null or within the range of children.',
+       ),
+       super(fit: sizing);
+
+  /// The index of the child to show.
+  final int? index;
+
+  bool _debugCheckHasDirectionality(BuildContext context) {
+    if (alignment is AlignmentDirectional && textDirection == null) {
+      assert(
+        debugCheckHasDirectionality(
+          context,
+          why: "to resolve the 'alignment' argument",
+          hint: alignment == AlignmentDirectional.topStart
+              ? "The default value for 'alignment' is AlignmentDirectional.topStart, which requires a text direction."
+              : null,
+          alternative:
+              "Instead of providing a Directionality widget, another solution would be passing a non-directional 'alignment', or an explicit 'textDirection', to the $runtimeType.",
+        ),
+      );
+    }
+    return true;
+  }
+
+  @override
+  RenderIndexedStack createRenderObject(BuildContext context) {
+    assert(_debugCheckHasDirectionality(context));
+    return RenderIndexedStack(
+      index: index,
+      fit: fit,
+      clipBehavior: clipBehavior,
+      alignment: alignment,
+      textDirection: textDirection ?? Directionality.maybeOf(context),
+    );
+  }
+
+  @override
+  void updateRenderObject(BuildContext context, RenderIndexedStack renderObject) {
+    assert(_debugCheckHasDirectionality(context));
+    renderObject
+      ..index = index
+      ..fit = fit
+      ..clipBehavior = clipBehavior
+      ..alignment = alignment
+      ..textDirection = textDirection ?? Directionality.maybeOf(context);
+  }
+
+  @override
+  MultiChildRenderObjectElement createElement() {
+    return _IndexedStackElement(this);
+  }
+}
+
+class _IndexedStackElement extends MultiChildRenderObjectElement {
+  _IndexedStackElement(_RawIndexedStack super.widget);
+
+  @override
+  _RawIndexedStack get widget => super.widget as _RawIndexedStack;
+
+  @override
+  void debugVisitOnstageChildren(ElementVisitor visitor) {
+    final int? index = widget.index;
+    // If the index is null, no child is onstage. Otherwise, only the child at
+    // the selected index is.
+    if (index != null && children.isNotEmpty) {
+      visitor(children.elementAt(index));
+    }
+  }
+}
 
 /// Whether to show or hide a child.
 ///
@@ -250,15 +435,15 @@ class Visibility extends StatelessWidget {
     var isVisible = true;
     var ancestorContext = context;
     InheritedElement? ancestor = ancestorContext
-        .getElementForInheritedWidgetOfExactType<VisibilityScope>();
+        .getElementForInheritedWidgetOfExactType<_VisibilityScope>();
     while (isVisible && ancestor != null) {
-      final scope = context.dependOnInheritedElement(ancestor) as VisibilityScope;
+      final scope = context.dependOnInheritedElement(ancestor) as _VisibilityScope;
       isVisible = scope.isVisible;
       ancestor.visitAncestorElements((Element parent) {
         ancestorContext = parent;
         return false;
       });
-      ancestor = ancestorContext.getElementForInheritedWidgetOfExactType<VisibilityScope>();
+      ancestor = ancestorContext.getElementForInheritedWidgetOfExactType<_VisibilityScope>();
     }
     return isVisible;
   }
@@ -287,7 +472,7 @@ class Visibility extends StatelessWidget {
         result = visible ? child : replacement;
       }
     }
-    return VisibilityScope(isVisible: visible, child: result);
+    return _VisibilityScope(isVisible: visible, child: result);
   }
 
   @override
@@ -314,18 +499,15 @@ class Visibility extends StatelessWidget {
 
 /// Inherited widget that allows descendants to find their visibility status.
 ///
-/// Used internally by [Visibility] and [IndexedStack] to propagate visibility
-/// information to descendants.
-@internal
-class VisibilityScope extends InheritedWidget {
-  /// Creates a [VisibilityScope].
-  const VisibilityScope({super.key, required this.isVisible, required super.child});
+/// Used by [Visibility] and [IndexedStack] to propagate visibility information
+/// to descendants.
+class _VisibilityScope extends InheritedWidget {
+  const _VisibilityScope({required this.isVisible, required super.child});
 
-  /// Whether the subtree is visible.
   final bool isVisible;
 
   @override
-  bool updateShouldNotify(VisibilityScope old) {
+  bool updateShouldNotify(_VisibilityScope old) {
     return isVisible != old.isVisible;
   }
 }
