+++ packages/flutter/lib/widgets.dart
@@ -73,6 +73,7 @@ export 'src/widgets/image.dart';
 export 'src/widgets/image_filter.dart';
 export 'src/widgets/image_icon.dart';
 export 'src/widgets/implicit_animations.dart';
+export 'src/widgets/indexed_stack.dart';
 export 'src/widgets/inherited_model.dart';
 export 'src/widgets/inherited_notifier.dart';
 export 'src/widgets/inherited_theme.dart';
@@ -181,7 +182,6 @@ export 'src/widgets/unique_widget.dart';
 export 'src/widgets/value_listenable_builder.dart';
 export 'src/widgets/view.dart';
 export 'src/widgets/viewport.dart';
-export 'src/widgets/visibility.dart' hide VisibilityScope;
 export 'src/widgets/widget_inspector.dart';
 export 'src/widgets/widget_span.dart';
 export 'src/widgets/widget_state.dart';
